# graphql-schema/src/test/resources/books-swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Expose resources via REST",
    "version": "0.0.1",
    "title": "REST service"
  },
  "host": "localhost:8081",
  "basePath": "/",
  "tags": [
    {
      "name": "books-controller",
      "description": "Books Controller"
    }
  ],
  "paths": {
    "/books/": {
      "get": {
        "tags": [
          "API Books"
        ],
        "summary": "Returns a list of all books",
        "operationId": "booksUsingGET",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/BookDto"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      }
    },
    "/books/library/{id}": {
      "get": {
        "tags": [
          "API Books"
        ],
        "summary": "Returns book in library by id",
        "operationId": "booksByLibraryIdUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/BookDto"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      }
    },
    "/books/{id}": {
      "get": {
        "tags": [
          "API Books"
        ],
        "summary": "Returns book by id",
        "operationId": "bookByIdUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BookDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      }
    }
  },
  "definitions": {
    "AuthorDto": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "lastName": {
          "type": "string"
        }
      },
      "title": "AuthorDto"
    },
    "BookDto": {
      "type": "object",
      "properties": {
        "author": {
          "$ref": "#/definitions/AuthorDto"
        },
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "pageCount": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "BookDto"
    }
  }
}